requirements.txt
#to install this requirements use the below command
# pip install -r requirements.txt


# Flask: Core framework
Flask

# Flask-SQLAlchemy: ORM for database interactions
Flask-SQLAlchemy

# PyMySQL: MySQL driver
PyMySQL

# python-dotenv: For environment variable management
python-dotenv

# Flask-WTF: For form handling and validation (Optional)
Flask-WTF

# Flask-Uploads: For managing file uploads (Optional)
Flask-Uploads

# Flask-Mail: For sending emails from your application (Optional)
Flask-Mail

# fpdf: For PDF generation
fpdf

# Export environment variables
export FLASK_APP=app
export FLASK_ENV=development  # or production

# Install dependencies
pip install -r requirements.txt

# Create database tables
python app.py

# Run the Flask app
python app.py
